RT4Sentinel/requirements.txt
fastapi
fastapi[standard]
python-dotenv
requests
sqlalchemy
uvicorn